requirements.txt
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
pyyaml==6.0.2pip


